Neil

+++ app/src/main/java/com/example/seg2105_termprojectcourseenrollmentapp/AdminHomePage.java
@@ -21,9 +21,9 @@ public class AdminHomePage extends AppCompatActivity {
 
     private AppBarConfiguration appBarConfiguration;
     private ActivityHomePageBinding binding;
-    private TextView wlcmAdminmessage;
-    private Button createClass, deleteClass, edtClass, searchUsers;
-    private EditText searchforUsers, newClassCode, newClassName;
+    private TextView wlcmAdminmessage, errorMessage;
+    private Button createClass, deleteClass, edtClass, searchUsers, searchCourse, returnToHome;
+    private EditText searchforUsers, oldClassCode, oldClassName, newClassCode, newClassName;
     private DBHelper db;
 
 
@@ -37,31 +37,81 @@ public class AdminHomePage extends AppCompatActivity {
         edtClass = (Button) findViewById(R.id.editClass);
         searchUsers = (Button) findViewById(R.id.searchUsers);
         searchforUsers = (EditText) findViewById(R.id.searchBar);
-        newClassCode = (EditText) findViewById(R.id.newCourseCode);
-        newClassName = (EditText) findViewById(R.id.newNameBar);
+        oldClassCode = (EditText) findViewById(R.id.newCourseCode);
+        oldClassName = (EditText) findViewById(R.id.newNameBar);
+        newClassName = (EditText) findViewById(R.id.newName);
+        newClassCode = (EditText) findViewById(R.id.newCode);
+        errorMessage = (TextView) findViewById(R.id.errorCode);
+        searchCourse = (Button) findViewById(R.id.searchCourses);
+        returnToHome = (Button) findViewById(R.id.returnToHome);
         db = new DBHelper(this);
         Bundle extras = getIntent().getExtras();
         String[] name = db.getName(extras.getString("username"));
         wlcmAdminmessage.setText("Welcome "+name[0]+" "+name[1]+"! you are logged in as admin");
+
+        createClass.setOnClickListener(this::onClick);
+        deleteClass.setOnClickListener(this::onClick);
+        edtClass.setOnClickListener(this::onClick);
+        searchUsers.setOnClickListener(this::onClick);
+        searchCourse.setOnClickListener(this::onClick);
+        returnToHome.setOnClickListener(this::onClick);
     }
 
     public void onClick(View view) {
+        String oldcode;
+        String oldName;
         switch(view.getId()){
             case R.id.createClass:
-                if(validInput((newClassCode).toString(),(newClassName).toString())){
-                    createCourse((newClassCode).toString(),(newClassName).toString());
+                if(validInput((oldClassCode).toString(),(oldClassName).toString())){
+                    createCourse((oldClassCode).toString(),(oldClassName).toString());
                 }
-            case R.id.deleteClass:
-                if(validInput((newClassCode).toString(),(newClassName).toString())) {
-                    deleteCourse((newClassCode).toString(),(newClassName).toString());
-                }
-            case R.id.editClass:
-                if(validInput((newClassCode).toString(),(newClassName).toString())){
-                    editCourse(newClassCode.toString(),newClassName.toString());
-                }
+
+                break;
+
             case R.id.searchUsers:
-                deleteUsers(searchforUsers.getText().toString());
-                
+                if (validInput(searchforUsers.getText().toString())) {
+                    deleteUsers(searchforUsers.getText().toString());
+                }
+                else {
+                    errorMessage.setText("");
+                    oldClassCode.setText("");
+                    oldClassName.setText("");
+                }
+                break;
+            case R.id.searchCourses:
+                oldcode = oldClassCode.getText().toString();
+                oldName = oldClassName.getText().toString();
+                if (validInput(oldcode, oldName)) {
+                    errorMessage.setText("");
+                    searchforUsers.setVisibility(view.GONE);
+                    edtClass.setVisibility(view.VISIBLE);
+                    deleteClass.setVisibility(view.VISIBLE);
+                    returnToHome.setVisibility(view.VISIBLE);
+                    searchUsers.setVisibility(view.GONE);
+                    createClass.setVisibility(view.GONE);
+                    newClassCode.setVisibility(view.VISIBLE);
+                    newClassName.setVisibility(view.VISIBLE);
+                }
+                break;
+            case R.id.returnToHome:
+                searchforUsers.setVisibility(view.VISIBLE);
+                searchUsers.setVisibility(view.VISIBLE);
+                createClass.setVisibility(view.VISIBLE);
+                edtClass.setVisibility(view.GONE);
+                deleteClass.setVisibility(view.GONE);
+                returnToHome.setVisibility(view.GONE);
+                newClassCode.setVisibility(view.GONE);
+                newClassName.setVisibility(view.GONE);
+                oldClassName.setText("");
+                oldClassCode.setText("");
+                break;
+            case R.id.editClass:
+                if (validInput(newClassCode.getText().toString(), newClassName.getText().toString())) {
+                    editCourse(newClassCode.getText().toString(), newClassName.getText().toString(), oldClassCode.getText().toString(), oldClassName.getText().toString());
+                }
+
+                break;
+ 
         }
 
 
@@ -73,8 +123,8 @@ public class AdminHomePage extends AppCompatActivity {
                 || super.onSupportNavigateUp();
     }
 
-    private void editCourse(String classCode, String className){
-
+    private void editCourse(String classCode, String className, String oldCode, String oldName){
+        db.editCourse(classCode, className, oldCode, oldName);
     }
     /**
      *
@@ -96,7 +146,17 @@ public class AdminHomePage extends AppCompatActivity {
     }
 
     private void deleteUsers(String username){
-        db.removeUser(username);
+
+        if (db.userExists(searchforUsers.getText().toString())) {
+            db.removeUser(username);
+        }
+        else {
+            errorMessage.setText("user doesn't exist");
+            searchUsers.setText("");
+            oldClassCode.setText("");
+            oldClassName.setText("");
+        }
+
     }
     /**
      *
@@ -107,7 +167,7 @@ public class AdminHomePage extends AppCompatActivity {
     private boolean validInput(String input1){
         boolean returnValue = true;
         if (input1.equals("")) {
-            wlcmAdminmessage.setText("Please enter a valid username or password");
+            errorMessage.setText("Please enter a valid username or password");
             searchforUsers.setText("");
             returnValue=false;
         }
@@ -124,9 +184,9 @@ public class AdminHomePage extends AppCompatActivity {
     private boolean validInput(String input1, String input2){
         boolean valReturn = true;
         if (input1.equals("")||input2.equals("")) {
-            wlcmAdminmessage.setText("Please enter a valid username or password");
-            newClassCode.setText("");
-            newClassName.setText("");
+            errorMessage.setText("Please enter a valid username or password");
+            oldClassCode.setText("");
+            oldClassName.setText("");
             valReturn=false;
         }
         return valReturn;
